+++ lib/glimmer/libui/transformable.rb
@@ -19,13 +19,24 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
-require 'glimmer/libui/control_proxy'
-
 module Glimmer
   module LibUI
-    # Includes transform property to controls or shapes that can apply a transform
-    # during their draw operation
+    # This is meant to be prepended (not included) because it changes behavior of instance draw method
+    # Adds transform property to controls/shapes
+    # Automatically applies transform property at the beginning of draw operation and undoes it at the end
+    # Stacks up with Parent module (must include Parent beforehand)
+    # Expects transformable to implement redraw method
     module Transformable
+      def post_initialize_child(child, add_child: true)
+        if child.is_a?(MatrixProxy)
+          super(child, add_child: false)
+          self.transform = child if child.keyword == 'transform'
+        else
+          super(child, add_child: add_child)
+        end
+      end
+    
+      # Returns transform or sets it. Expects transformable to implement redraw method (delegating work to area).
       def transform(matrix = nil)
         if matrix.nil?
           @transform
@@ -36,6 +47,26 @@ module Glimmer
       end
       alias transform= transform
       alias set_transform transform
+      
+      # Apply transform matrix to coordinate system
+      def apply_transform(area_draw_params)
+        ::LibUI.draw_transform(area_draw_params[:context], @transform.libui) unless @transform.nil?
+      end
+      
+      # Inverse of apply_transform (applies inverse transformation to undo initial transformation)
+      def undo_transform(area_draw_params)
+        unless @transform.nil?
+          inverse_transform = @transform.clone
+          inverse_transform.invert
+          ::LibUI.draw_transform(area_draw_params[:context], inverse_transform.libui)
+        end
+      end
+      
+      def draw(area_draw_params)
+        apply_transform(area_draw_params)
+        super(area_draw_params)
+        undo_transform(area_draw_params)
+      end
     end
   end
 end